# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Nimbus control plane API
  version: 0.1.0
  description: Tenant-scoped HTTP contracts for the first durable task slice.
servers:
  - url: http://localhost:3000
paths:
  /api/tasks:
    post:
      operationId: createTask
      summary: Create an idempotent task in an authorized repository
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/CreateTask"
      responses:
        "303":
          description: Task created or replayed
          headers:
            Location:
              schema: { type: string }
        "400": { description: Invalid request }
        "401": { description: Authentication required }
        "404": { description: Repository is not available to the tenant }
        "409": { description: Idempotency conflict }
  /api/tasks/{taskId}/events:
    get:
      operationId: streamTaskEvents
      summary: Replay and stream authorized durable task events
      parameters:
        - in: path
          name: taskId
          required: true
          schema: { type: string }
        - in: header
          name: Last-Event-ID
          schema: { type: integer, minimum: 0 }
      responses:
        "200":
          description: Server-sent event stream ordered by durable sequence
          content:
            text/event-stream:
              schema: { type: string }
        "401": { description: Authentication required }
        "404": { description: Task is not visible to the tenant }
components:
  schemas:
    CreateTask:
      type: object
      required: [repositoryId, objective, model, idempotencyKey]
      properties:
        repositoryId: { type: string, minLength: 12 }
        objective: { type: string, minLength: 10, maxLength: 8000 }
        model:
          type: string
          description: Model ID from the connected account's current Codex app-server catalog.
        idempotencyKey: { type: string, minLength: 12, maxLength: 200 }